Bugfix for RegEx's in Zxcvbn.PasswordScoring
StartUpper gives you some wrong answers :
ScoRpi0ns is not a match but should be
ROSEBUD is not a match but should be

EndUpper gives you some wrong answers :
ROSEBUD is not a match but should be
rosebuD is not a match but should be
R0$38uD is not a match but should be
VANDERMARK is not a match but should be

Another problem is the lag of support for non English chars. Eg the Danish ones: ÆØÅæøå

All fixed!